fix casting

+++ src/libtorchaudio/forced_align/cpu/compute.cpp
@@ -129,8 +129,8 @@ void forced_align_impl(
     auto lbl_idx = ltrIdx % 2 == 0 ? blank : targets_a[batchIndex][ltrIdx / 2];
     paths_a[batchIndex][t] = lbl_idx;
     // Calculate backPtr value from bits
-    auto backPtr_idx = backPtr_seek[std::max(t - 1, 0)] + ltr_idx -
-        backPtr_offset[std::max(t - 1, 0)];
+    auto backPtr_idx = backPtr_seek[std::max(t - 1, static_cast<long int>(0))] +
+        ltr_idx - backPtr_offset[std::max(t - 1, static_cast<long int>(0))];
     ltrIdx -= (backPtrBit1[backPtr_idx] << 1) | backPtrBit0[backPtr_idx];
   }
 }
